requirements.txt
flask
flask-sqlalchemy
flask-login
werkzeug
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
python-dotenv
gunicorn
apscheduler
